MSD: Extract shared staging site logic into useStagingSite hook
The environment-switcher and environment-switcher-v2 components shared
~380 lines of identical staging site orchestration logic (queries,
mutations, effects, and the creation handler). This extracts that
logic into a reusable useStagingSite hook, reducing both components
to thin render wrappers.

Closes DOTMSD-1155
